LayoutTests/accessibility/roles-exposed.html

<!DOCTYPE HTML>
<html>
<body>
<script src="../resources/js-test-pre.js"></script>

<!-- ==================================================================================================== -->
<!-- HTML elements in alphabetical order, excepting the need for nesting (e.g. <li> is with its <ol> and <ul> parents) -->
<!-- ==================================================================================================== -->
<a data-platform="atk,mac" data-role="AXLink" data-subrole="" data-roledescription="link" href="#" data-note="[href]" class="ex">X</a>
<!-- [ATK] Skipped: not focusable --><a data-platform="mac" data-role="" data-subrole="" data-roledescription="" data-note=":not([href])" class="ex">X</a>
<!-- [ATK] Skipped: not focusable --><abbr data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</abbr>
<!-- http://webkit.org/b/83989 --><address data-platform="atk,mac" data-role-atk="AXLandmarkContentInfo" data-role-mac="AXGroup" data-subrole="AXLandmarkContentInfo" data-roledescription="content information" class="ex">X</address>
<!-- skipped <area> -->
<article data-platform="atk,mac" data-role-atk="AXArticle" data-role-mac="AXGroup" data-subrole="AXDocumentArticle" data-roledescription="article" class="ex">X</article>
<aside data-platform="atk,mac" data-role-atk="AXLandmarkComplementary" data-role-mac="AXGroup" data-subrole="AXLandmarkComplementary" data-roledescription="complementary" class="ex">X</aside>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><audio data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</audio>
<!-- [ATK] Skipped: not focusable --><b data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</b>
<!-- [ATK] Skipped: not focusable --><bdo data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</bdo>
<blockquote data-platform="atk,mac" data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</blockquote>
<button data-platform="atk,mac" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex">X</button>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><canvas data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</canvas>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><cite data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</cite>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><code data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</code>
<!-- skipped <command> -->
<!-- skipped <datalist> -->
<!-- [ATK] Skipped: not focusable --><del data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</del>
<!-- skipped <details/summary> http://webkit.org/b/108979 -->
<!-- [ATK] Skipped: not focusable --><dfn data-platform="mac" data-role="AXGroup" data-subrole="AXDefinition" data-roledescription="definition" class="ex">X</dfn>
<dl data-platform="atk,mac" data-role-atk="AXDescriptionList" data-role-mac="AXList" data-subrole="AXDescriptionList" data-alternatesubrole="AXDefinitionList" data-roledescription="description list" class="ex">
    <dt data-platform="atk,mac" data-role-atk="AXDescriptionTerm" data-role-mac="AXGroup" data-subrole="AXTerm" data-roledescription="term" class="ex">X</dt>
    <dd data-platform="atk,mac" data-role-atk="AXDescriptionValue" data-role-mac="AXGroup" data-subrole="AXDescription" data-roledescription="description" class="ex">X</dd>
</dl>
<div data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<!-- [ATK] Skipped: not focusable --><em data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</em>
<!-- skipped <embed> -->
<fieldset data-platform="atk,mac" data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">
    <legend data-platform="atk,mac" data-role="" data-role-atk="AXLabel" data-subrole="" data-roledescription="" class="ex">X</legend>
</fieldset>
<!-- skipped figure/figcaption http://webkit.org/b/108996 -->
<!-- [ATK] Skipped: not focusable --><footer data-platform="mac" data-role="AXGroup" data-subrole="AXLandmarkContentInfo" data-roledescription="footer" class="ex">X</footer>
<form data-platform="atk,mac" data-role-atk="AXForm" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</form>
<!-- skipped <frame> -->
<!-- skipped <frameset> -->
<!-- skipped <head> -->
<!-- http://webkit.org/b/109013 --><header data-platform="atk,mac" data-role-atk="AXLandmarkBanner" data-role-mac="AXGroup" data-subrole="AXLandmarkBanner" data-roledescription="banner" class="ex">X</header>
<hgroup data-platform="atk,mac" data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</hgroup>
<h1 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h1>
<h2 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h2>
<h3 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h3>
<h4 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h4>
<h5 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h5>
<h6 data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</h6>
<hr data-platform="atk,mac" data-role-mac="AXSplitter" data-role-atk="AXSeparator" data-subrole="AXContentSeparator" data-roledescription="separator" class="ex">
<!-- skipped <html> -->
<!-- [ATK] Skipped: not focusable --><i data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</i>
<!-- skipped <iframe> -->
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><img data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex" data-note=":not([src]):not([alt])">
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><img data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex" alt="" data-note="[alt='']">
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><img data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex" src="foo.png" data-note="[src]:not([alt])">
<img data-platform="atk,mac" data-role="AXImage" data-subrole="" data-roledescription="image" class="ex" alt="X" data-note="[alt='X']">
<input data-platform="atk,mac" type="button" value="X" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex" data-note="[type='button']">
<!-- skipped <menu><input type="button"></menu> -->
<!-- [ATK] Wrong role (webkit.org/b/125493) --><!-- http://webkit.org/b/109017 --><input data-platform="mac" type="color" value="X" data-role="AXColorWell" data-subrole="" data-roledescription="color well" class="ex" data-note="[type='color']">
<input data-platform="atk,mac" type="checkbox" data-role="AXCheckBox" data-subrole="" data-roledescription="check box" class="ex" data-note="[type='checkbox']">
<!-- skipped <menu><input type="checkbox"></menu> -->
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="date" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='date']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="datetime" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='datetime']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="datetime-local" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='datetime-local']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="email" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='email']">
<input data-platform="atk,mac" type="file" data-role="AXButton" data-subrole="AXFileUploadButton" data-roledescription="file upload button" class="ex" data-note="[type='file']">
<!-- [ATK] Skipped: not focusable --><input data-platform="mac" type="hidden" data-role="" data-subrole="" data-roledescription="" class="ex" data-note="[type='hidden']">
<input data-platform="atk,mac" type="image" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex" data-note="[type='image']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="month" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='month']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="number" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='number']">
<input data-platform="atk,mac" type="password" value="X" data-role-atk="AXPasswordField" data-role-mac="AXTextField" data-subrole="AXSecureTextField" data-roledescription="secure text field" class="ex" data-note="[type='password']">
<input data-platform="atk,mac" type="radio" data-role="AXRadioButton" data-subrole="" data-roledescription="radio button" class="ex" data-note="[type='radio']">
<!-- skipped <menu><input type="radio"></menu> -->
<input data-platform="atk,mac" type="range" data-role="AXSlider" data-subrole="" data-roledescription="slider" class="ex" data-note="[type='range']">
<input data-platform="atk,mac" type="reset" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex" data-note="[type='reset']">
<input data-platform="atk,mac" type="search" value="X" data-role="AXTextField" data-subrole="AXSearchField" data-roledescription="search text field" class="ex" data-note="[type='search']">
<input data-platform="atk,mac" type="submit" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex" data-note="[type='submit']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="tel" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='tel']">
<input data-platform="atk,mac" type="text" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='text']">
<!-- skipped <input type="text"> with suggestions source element: http://www.w3.org/html/wg/drafts/html/master/forms.html#concept-input-list -->
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="time" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='time']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="url" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='url']">
<!-- http://webkit.org/b/109017 --><input data-platform="atk,mac" type="week" value="X" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex" data-note="[type='week']">
<!-- [ATK] Skipped: not focusable --><ins data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</ins>
<!-- skipped <link> -->
<!-- skipped <map> -->
<!-- [ATK] Skipped: not focusable --><!-- http://webkit.org/b/109018 --><mark data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</mark>
<math data-platform="atk,mac" data-role-atk="AXMath" data-role-mac="AXGroup" data-subrole="AXDocumentMath" data-roledescription="math" class="ex">X</math>
<!-- skipped <menu> -->
<!-- skipped <meta> -->
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><meter data-platform="mac" data-role="AXProgressIndicator" data-subrole="" data-roledescription="progress indicator" class="ex" value="0.75">X</meter>
<nav data-platform="atk,mac" data-role-atk="AXLandmarkNavigation" data-role-mac="AXGroup" data-subrole="AXLandmarkNavigation" data-roledescription="navigation" class="ex">X</nav>
<!-- skipped <noscript> -->
<!-- skipped <object> -->
<ol data-platform="atk,mac" data-role="AXList" data-subrole="AXContentList" data-roledescription="content list" class="ex">
    <li data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</li>
</ol>
<!-- skipped <optgroup> -->
<!-- skipped <option> -->
<!-- skipped <output> -->
<p data-platform="atk,mac" data-role-atk="AXParagraph" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</p>
<!-- skipped <param> -->
<pre data-platform="atk,mac" data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</pre>
<progress data-platform="atk,mac" data-role="AXProgressIndicator" data-subrole="" data-roledescription="progress indicator" class="ex" value="0.75">X</progress>
<!-- [ATK] Skipped: not focusable --><q data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</q>
<!-- skipped <ruby/rp/rt> -->
<!-- [ATK] Skipped: not focusable --><s data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</s>
<!-- [ATK] Skipped: not focusable --><samp data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</samp>
<!-- skipped <script> -->
<!-- [ATK] Wrong role (webkit.org/b/125493) --><!-- http://webkit.org/b/109024 --><section data-platform="mac" data-role="AXGroup" data-subrole="AXDocumentRegion" data-roledescription="region" class="ex">X</section>
<select data-platform="atk,mac" data-role-atk="AXComboBox" data-role-mac="AXPopUpButton" data-subrole="" data-roledescription="pop up button" class="ex" data-note=":not([multiple])">
    <!-- [ATK] Object not exposed (webkit.org/b/125491) --><option data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</option>
    <!-- [ATK] Object not exposed (webkit.org/b/125491) --><optgroup data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex" label="more">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><option data-platform="mac">X</option>
    </optgroup>
</select>
<select data-platform="atk,mac" data-role="AXList" data-subrole="" data-roledescription="list" class="ex" multiple data-note="[multiple]">
    <option data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXStaticText" data-subrole="" data-roledescription="text" class="ex">X</option>
    <optgroup data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXStaticText" data-subrole="" data-roledescription="text" class="ex" label="more">
        <option>Y</option>
        <option>Z</option>
    </optgroup>
</select>
<!-- [ATK] Skipped: not focusable --><small data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</small>
<!-- skipped <source> -->
<!-- [ATK] Skipped: not focusable --><span data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</span>
<!-- [ATK] Skipped: not focusable --><strong data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</strong>
<!-- skipped <style> -->
<!-- [ATK] Skipped: not focusable --><sub data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</sub>
<!-- [ATK] Skipped: not focusable --><sup data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</sup>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><svg data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</svg>

<table data-platform="atk,mac" data-role="AXTable" data-subrole="" data-roledescription="table" class="ex">
    <!-- [ATK] Skipped: not to be exposed --><caption data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</caption>
    <!-- Not to be exposed in ATK--><thead data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><tr data-platform="mac" data-role="AXRow" data-subrole="" data-roledescription="row" class="ex">
            <!-- Need separate test to verify colheader/rowheader -->
            <th data-platform="atk,mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</th>
        </tr>
    </thead>
    <!-- [ATK] Skipped: not to be exposed --><tbody data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><tr data-platform="mac" data-role="AXRow" data-subrole="" data-roledescription="row" class="ex">
            <td data-platform="atk,mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</td>
        </tr>
    </tbody>
    <!-- [ATK] Skipped: not to be exposed --><tfoot data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><tr data-platform="mac" data-role="AXRow" data-subrole="" data-roledescription="row" class="ex">
            <!-- Need separate test to verify colheader/rowheader -->
            <th data-platform="atk,mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</th>
        </tr>
    </tfoot>
</table>
<textarea data-platform="atk,mac" data-role-atk="AXTextField" data-role-mac="AXTextArea" data-subrole="" data-roledescription="text entry area" class="ex">X</textarea>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><time data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</time>
<!-- skipped <title> -->
<ul data-platform="atk,mac" data-role="AXList" data-subrole="AXContentList" data-roledescription="content list" class="ex">
    <li data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</li>
</ul>
<!-- [ATK] Skipped: not focusable --><var data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</var>
<!-- skipped <video> -->
<!-- [ATK] Skipped: not focusable --><wbr data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</wbr>
<!-- Todo: I left off adding HTML5 elements here at previous element: https://dvcs.w3.org/hg/html-api-map/raw-file/default/Overview.html -->



<!-- ==================================================================================================== -->
<!-- Abstract ARIA roles in alphabetical order; only generic AXGroup role should be exposed on abstract roles -->
<!-- ==================================================================================================== -->
<div role="command"     data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="composite"   data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="input"       data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="landmark"    data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="range"       data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="roletype"    data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="section"     data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="sectionhead" data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="select"      data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="structure"   data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="widget"      data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="window"      data-platform="atk,mac" data-role-atk="AXSection" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>


<!-- ==================================================================================================== -->
<!-- Non-abstract ARIA roles in alphabetical order, excepting the need for nesting (e.g. row is with its grid parent) -->
<!-- ==================================================================================================== -->
<div role="alert"                    data-platform="atk,mac" data-role-atk="AXAlert" data-role-mac="AXGroup" data-subrole="AXApplicationAlert" data-roledescription="alert" class="ex">X</div>
<div role="alertdialog"              data-platform="atk,mac" data-role-atk="AXAlert" data-role-mac="AXGroup" data-subrole="AXApplicationAlertDialog" data-roledescription="alert dialog" class="ex">X</div>
<div role="application"              data-platform="atk,mac" data-role-atk="AXEmbedded" data-role-mac="AXGroup" data-subrole="AXLandmarkApplication" data-roledescription="application" class="ex">X</div>
<div role="article"                  data-platform="atk,mac" data-role-atk="AXArticle" data-role-mac="AXGroup" data-subrole="AXDocumentArticle" data-roledescription="article" class="ex">X</div>
<div role="banner"                   data-platform="atk,mac" data-role-atk="AXLandmarkBanner" data-role-mac="AXGroup" data-subrole="AXLandmarkBanner" data-roledescription="banner" class="ex">X</div>
<div role="button"                   data-platform="atk,mac" data-role="AXButton" data-subrole="" data-roledescription="button" class="ex">X</div>
<div role="checkbox"                 data-platform="atk,mac" data-role="AXCheckBox" data-subrole="" data-roledescription="check box" class="ex">X</div>
<div role="combobox"                 data-platform="atk,mac" data-role="AXComboBox" data-subrole="" data-roledescription="combo box" class="ex">X</div>
<div role="complementary"            data-platform="atk,mac" data-role-atk="AXLandmarkComplementary" data-role-mac="AXGroup" data-subrole="AXLandmarkComplementary" data-roledescription="complementary" class="ex">X</div>
<div role="contentinfo"              data-platform="atk,mac" data-role-atk="AXLandmarkContentInfo" data-role-mac="AXGroup" data-subrole="AXLandmarkContentInfo" data-roledescription="content information" class="ex">X</div>
<div role="definition"               data-platform="atk,mac" data-role-atk="AXDefinition" data-role-mac="AXGroup" data-subrole="AXDefinition" data-roledescription="definition" class="ex">X</div>
<div role="dialog"                   data-platform="atk,mac" data-role-atk="AXDialog" data-role-mac="AXGroup" data-subrole="AXApplicationDialog" data-roledescription="dialog" class="ex">X</div>
<div role="directory"                data-platform="atk,mac" data-role="AXList" data-subrole="AXContentList" data-roledescription="content list" class="ex">X</div>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><div data-platform="mac" role="document" data-role="AXGroup" data-subrole="AXDocument" data-roledescription="document" class="ex">X</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div data-platform="mac" role="form"     data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div data-platform="atk,mac" role="grid"                     data-role="AXTable" data-subrole="" data-roledescription="table" class="ex">
    <!-- [ATK] Object not exposed (webkit.org/b/125491) --><div data-platform="mac" role="rowgroup"             data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><div data-platform="mac" role="row"              data-role="AXRow" data-subrole="" data-roledescription="row" class="ex">
            <!-- Note: diff between rowheader, columnheader, and gridcell is in cross-reference from grid; need additional non-role verification in another test. -->
            <!-- [ATK] Wrong role (webkit.org/b/125493) --><div data-platform="mac" role="rowheader"    data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
            <!-- [ATK] Wrong role (webkit.org/b/125493) --><div data-platform="mac" role="columnheader" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
            <div data-platform="atk,mac" role="gridcell"     data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
        </div>
    </div>
</div>
<div role="group"                    data-platform="atk,mac" data-role="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
<div role="heading"                  data-platform="atk,mac" data-role="AXHeading" data-subrole="" data-roledescription="heading" class="ex">X</div>
<div role="img"                      data-platform="atk,mac" data-role="AXImage" data-subrole="" data-roledescription="image" class="ex">X</div>
<div role="link"                     data-platform="atk,mac" data-role="AXLink" data-subrole="" data-roledescription="link" class="ex">X</div>
<div role="list"                     data-platform="atk,mac" data-role="AXList" data-subrole="AXContentList" data-roledescription="content list" class="ex">
    <div role="listitem"             data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXGroup" data-subrole="" data-roledescription="group" class="ex">X</div>
</div>
<!-- [ATK] Object not exposed (webkit.org/b/125491) --><div role="listbox" data-platform="mac" data-role="AXList" data-subrole="" data-roledescription="list" class="ex">
    <div role="option"               data-platform="atk,mac" data-role-atk="AXListItem" data-role-mac="AXStaticText" data-subrole="" data-roledescription="text" class="ex">X</div>
</div>
<div role="log"                      data-platform="atk,mac" data-role-atk="AXLog" data-role-mac="AXGroup" data-subrole="AXApplicationLog" data-roledescription="log" class="ex">X</div>
<div role="main"                     data-platform="atk,mac" data-role-atk="AXLandmarkMain" data-role-mac="AXGroup" data-subrole="AXLandmarkMain" data-roledescription="main" class="ex">X</div>
<div role="marquee"                  data-platform="atk,mac" data-role-atk="AXMarquee" data-role-mac="AXGroup" data-subrole="AXApplicationMarquee" data-roledescription="marquee" class="ex">X</div>
<div role="math"                     data-platform="atk,mac" data-role-atk="AXMath" data-role-mac="AXGroup" data-subrole="AXDocumentMath" data-roledescription="math" class="ex">X</div>
<div role="menu"                     data-platform="atk,mac" data-role="AXMenu" data-subrole="" data-roledescription="menu" class="ex">
    <!-- Note: diff between menuitem types is an attribute; need additional non-role verification in another test. -->
    <div role="menuitem"             data-platform="atk,mac" data-role="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
    <div role="menuitemcheckbox"     data-platform="atk,mac" data-role-atk="AXCheckMenuItem" data-role-mac="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
    <div role="menuitemradio"        data-platform="atk,mac" data-role-atk="AXRadioMenuItem" data-role-mac="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
</div>
<div role="menubar"                  data-platform="atk,mac" data-role="AXMenuBar" data-subrole="" data-roledescription="menu bar" class="ex">
    <!-- Note: diff between menuitem types is an attribute; need additional non-role verification in another test. -->
    <div role="menuitem"             data-platform="atk,mac" data-role="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
    <div role="menuitemcheckbox"     data-platform="atk,mac" data-role-atk="AXCheckMenuItem" data-role-mac="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
    <div role="menuitemradio"        data-platform="atk,mac" data-role-atk="AXRadioMenuItem" data-role-mac="AXMenuItem" data-subrole="" data-roledescription="menu item" class="ex">X</div>
</div>
<div role="navigation"               data-platform="atk,mac" data-role-atk="AXLandmarkNavigation" data-role-mac="AXGroup" data-subrole="AXLandmarkNavigation" data-roledescription="navigation" class="ex">X</div>
<div role="note"                     data-platform="atk,mac" data-role-atk="AXComment" data-role-mac="AXGroup" data-subrole="AXDocumentNote" data-roledescription="note" class="ex">X</div>
<!-- [ATK] Skipped: not to be exposed --><div role="presentation" data-platform="mac" data-platform="atk,mac" data-role="" data-subrole="" data-roledescription="" class="ex">X</div>
<div role="progressbar"              data-platform="atk,mac" data-role="AXProgressIndicator" data-subrole="" data-roledescription="progress indicator" class="ex">X</div>
<div role="radiogroup"               data-platform="atk,mac" data-role-atk="AXGroup" data-role-mac="AXRadioGroup" data-subrole="" data-roledescription="radio group" class="ex">
    <div role="radio"                data-platform="atk,mac" data-role="AXRadioButton" data-subrole="" data-roledescription="radio button" class="ex">X</div>
</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="region" data-platform="mac" data-role="AXGroup" data-subrole="AXDocumentRegion" data-roledescription="region" class="ex">X</div>
<div role="scrollbar"                data-platform="atk,mac" data-role="AXScrollBar" data-subrole="" data-roledescription="scroll bar" class="ex">X</div>
<div role="search"                   data-platform="atk,mac" data-role-atk="AXLandmarkSearch" data-role-mac="AXGroup" data-subrole="AXLandmarkSearch" data-roledescription="search" class="ex">X</div>
<div role="separator"                data-platform="atk,mac" data-role-atk="AXSeparator" data-role-mac="AXSplitter" data-subrole="" data-roledescription="splitter" class="ex">X</div>
<div role="slider"                   data-platform="atk,mac" data-role="AXSlider" data-subrole="" data-roledescription="slider" class="ex">X</div>
<div role="spinbutton"               data-platform="atk,mac" data-role-atk="AXSpinButton" data-role-mac="AXIncrementor" data-subrole="" data-roledescription="stepper" class="ex">X</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="status" data-platform="mac" data-role="AXGroup" data-subrole="AXApplicationStatus" data-roledescription="application status" class="ex">X</div>
<div role="tablist"                  data-platform="atk,mac" data-role="AXTabGroup" data-subrole="" data-roledescription="tab group" class="ex">
    <div role="tab"                  data-platform="atk,mac" data-role-atk="AXTab" data-role-mac="AXRadioButton" data-subrole="" data-roledescription="tab" class="ex">X</div>
</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="tabpanel" data-platform="mac" data-role="AXGroup" data-subrole="AXTabPanel" data-roledescription="tab panel" class="ex">X</div>
<div role="textbox"                  data-platform="atk,mac" data-role="AXTextField" data-subrole="" data-roledescription="text field" class="ex">X</div>
<div role="timer"                    data-platform="atk,mac" data-role-atk="AXTimer" data-role-mac="AXGroup" data-subrole="AXApplicationTimer" data-roledescription="timer" class="ex">X</div>
<div role="toolbar"                  data-platform="atk,mac" data-role="AXToolbar" data-subrole="" data-roledescription="toolbar" class="ex">X</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="tooltip"                  data-platform="mac" data-role="AXGroup" data-subrole="AXUserInterfaceTooltip" data-roledescription="tooltip" class="ex">X</div>
<!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="tree"                     data-platform="mac" data-role="AXOutline" data-subrole="" data-roledescription="outline" class="ex">
    <!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="treeitem"             data-platform="mac" data-role="AXRow" data-subrole="AXOutlineRow" data-roledescription="outline row" class="ex">X</div>
    <!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="group"                data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="treeitem"         data-platform="mac" data-role="AXRow" data-subrole="AXOutlineRow" data-roledescription="outline row" class="ex">X</div>
    </div>
</div>
<div role="treegrid"                 data-platform="atk,mac" data-role="AXTable" data-subrole="" data-roledescription="table" class="ex">
    <!-- [ATK] Object not exposed (webkit.org/b/125491) --><div role="rowgroup"             data-platform="mac" data-role="" data-subrole="" data-roledescription="" class="ex">
        <!-- [ATK] Object not exposed (webkit.org/b/125491) --><div role="row"              data-platform="mac" data-role="AXRow" data-subrole="" data-roledescription="row" class="ex">
            <!-- Note: diff between rowheader, columnheader, and gridcell is in cross-reference from grid; need additional non-role verification in another test. -->
            <!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="rowheader"    data-platform="mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
            <!-- [ATK] Wrong role (webkit.org/b/125493) --><div role="columnheader" data-platform="mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
            <div role="gridcell"     data-platform="atk,mac" data-role="AXCell" data-subrole="" data-roledescription="cell" class="ex">X</div>
        </div>
    </div>
</div>

<div id="console"></div>
<script>
if (window.testRunner && window.accessibilityController) {
    description("This tests that native elements and ARIA overrides result in the expected role, subrole and role description.")
    var currentPlatform = accessibilityController.platformName;
    var examples = document.querySelectorAll('.ex');
    var indent = "\n      "; // newline and six spaces
    var el, supportedPlatforms, ariaRole, axElement, role, subrole, roleDescription, output, expectedRole, expectation, matchedForPlatform, result;
    for (var i = 0, c = examples.length; i < c; i++) {
        el = examples[i];

        supportedPlatforms = el.getAttribute('data-platform');
        if (!supportedPlatforms || supportedPlatforms.indexOf(currentPlatform) == -1)
            continue;

        el.id = 'ex' + i;
        ariaRole = el.getAttribute('role');
        axElement = accessibilityController.accessibleElementById(el.id);
        if (axElement) {
            role = axElement.role;
            subrole = axElement.subrole;
            roleDescription = axElement.roleDescription;
        } else {
            role = 'AXRole: ';
            subrole = 'AXSubrole: ';
            roleDescription = 'AXRoleDescription: ';
        }

        output = el.tagName.toLowerCase() + (ariaRole?("[role="+ariaRole+"]"):'');
        if (el.getAttribute('data-note')) {
            output += el.getAttribute('data-note');
        }
        output += indent;
        output += role + indent;

        expectedRole = "";
        if (el.hasAttribute('data-role-' + currentPlatform)) {
            expectedRole = el.getAttribute('data-role-' + currentPlatform);
        } else if (el.hasAttribute('data-role')) {
            expectedRole = el.getAttribute('data-role');
        }

        expectation = expectedRole;
        matchedResults = (role.replace('AXRole: ','') == expectedRole)

        // In the Mac platform more things than just the role are tested so we need to
        // expand a bit the output and expected results, as well as add some extra checks.
        if (currentPlatform == "mac") {
            output += subrole + indent;
            output += roleDescription + indent;

            expectation += "/";
            expectation += el.getAttribute('data-subrole');
            if (el.getAttribute('data-alternatesubrole')) expectation += " or " + el.getAttribute('data-alternatesubrole');
            expectation += "/" + el.getAttribute('data-roledescription');

            matchedResults = matchedResults && (subrole.replace('AXSubrole: ','') == el.getAttribute('data-subrole') || subrole.replace('AXSubrole: ','') == el.getAttribute('data-alternatesubrole'));
            matchedResults = matchedResults && (roleDescription.replace('AXRoleDescription: ','') == el.getAttribute('data-roledescription'));
        }

        result = document.getElementById('console');
        if (matchedResults) {
            result.innerText += "PASS: " + output + "\n";
        } else {
            result.innerText += "FAIL: " + output + "Expected: " + expectation + "\n\n";
        }
    }

    // Once tests are complete, hide all the example markup.
    examples = document.querySelectorAll('.ex');
    for (var i = 0, c = examples.length; i < c; i++) {
        el = examples[i];
        el.style.display = 'none';
    }
}
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
